requirements.txt
fastapi==0.109.2
httptools==0.6.1
pip-chill==1.0.3
python-dotenv==1.0.1
pyyaml==6.0.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
